jitta-assistant/requirements.txt
duckduckgo-search==7.3.2
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==21.7
httpx==0.27.2
chromadb==0.5.5
sentence-transformers==2.7.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
